feat(README.md): add README file with title 'Money Saving Azure Logic Apps' to provide initial project information
feat: add applicationGatewayStopStart.json for managing Azure App Gateways

This commit adds a new JSON file that defines a workflow for managing Azure Application Gateways. The workflow includes actions for filtering App Gateways, stopping and starting them based on their operational state, and retrieving information about the App Gateways. This is useful for automating the management of App Gateways in an Azure environment.

feat: add new properties to the JSON schema for better data validation
chore: update default values for parameters to improve usability
feat: add recurrence trigger to automate the process on a weekly basis

feat: add virtualMachineStopStart.json to manage VM power state

This commit adds a new file, virtualMachineStopStart.json, which contains a script to manage the power state of Azure virtual machines. The script fetches a list of all resources tagged to be managed, filters out everything but the VMs, and then for each VM, it checks the power state. If the VM is deallocated, it starts the VM, otherwise, it deallocates the VM. This script is intended to automate the process of starting and stopping VMs, potentially saving costs and resources.